Move Containers reducer
