+++ plugins/services/src/js/components/forms/ContainerServiceFormSection.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react';
 import {Tooltip} from 'reactjs-components';
 
 import {FormReducer as ContainerReducer} from '../../reducers/serviceForm/Container';
-import {FormReducer as ContainersReducer} from '../../reducers/serviceForm/Containers';
+import {FormReducer as ContainersReducer} from '../../reducers/serviceForm/JSONReducers/Containers';
 import {findNestedPropertyInObject} from '../../../../../../src/js/utils/Util';
 import ContainerConstants from '../../constants/ContainerConstants';
 import FieldError from '../../../../../../src/js/components/form/FieldError';
