+++ plugins/services/src/js/reducers/JSONSingleContainerParser.js
@@ -14,7 +14,6 @@ import {JSONParser as network} from './serviceForm/Network';
 import {JSONParser as multiContainerNetwork} from './serviceForm/MultiContainerNetwork';
 import {JSONParser as volumeMounts} from './serviceForm/MultiContainerVolumes';
 import {simpleParser} from '../../../../../src/js/utils/ParserUtil';
-import {JSONParser as containers} from './serviceForm/Containers';
 
 module.exports = [
   simpleParser(['id']),
@@ -25,7 +24,6 @@ module.exports = [
   simpleParser(['gpus']),
   simpleParser(['cmd']),
   container,
-  containers,
   network,
   multiContainerNetwork,
   volumeMounts,
