+++ plugins/services/src/js/reducers/serviceForm/JSONReducers/__tests__/Containers-test.js
@@ -1,8 +1,8 @@
 const Containers = require('../Containers');
-const Batch = require('../../../../../../../src/js/structs/Batch');
-const Transaction = require('../../../../../../../src/js/structs/Transaction');
-const {ADD_ITEM, SET} = require('../../../../../../../src/js/constants/TransactionTypes');
-const {DEFAULT_POD_CONTAINER} = require('../../../constants/DefaultPod');
+const Batch = require('../../../../../../../../src/js/structs/Batch');
+const Transaction = require('../../../../../../../../src/js/structs/Transaction');
+const {ADD_ITEM, SET} = require('../../../../../../../../src/js/constants/TransactionTypes');
+const {DEFAULT_POD_CONTAINER} = require('../../../../constants/DefaultPod');
 
 describe('Containers', function () {
